test: move mocks and fixtures to Design Electric conftest.py
