+++ tests/models/design_electric/design_electric_integration_test.py
@@ -19,68 +19,6 @@ from ramstk.models import RAMSTKDesignElectricTable
 from ramstk.models.programdb import RAMSTKDesignElectric
 
 
-@pytest.fixture(scope="function")
-def test_attributes():
-    yield {
-        "revision_id": 1,
-        "hardware_id": 1,
-        "application_id": 0,
-        "area": 0.0,
-        "capacitance": 0.0,
-        "configuration_id": 0,
-        "construction_id": 0,
-        "contact_form_id": 0,
-        "contact_gauge": 0,
-        "contact_rating_id": 0,
-        "current_operating": 0.0,
-        "current_rated": 0.0,
-        "current_ratio": 0.0,
-        "environment_active_id": 0,
-        "environment_dormant_id": 0,
-        "family_id": 0,
-        "feature_size": 0.0,
-        "frequency_operating": 0.0,
-        "insert_id": 0,
-        "insulation_id": 0,
-        "manufacturing_id": 0,
-        "matching_id": 0,
-        "n_active_pins": 0,
-        "n_circuit_planes": 1,
-        "n_cycles": 0,
-        "n_elements": 0,
-        "n_hand_soldered": 0,
-        "n_wave_soldered": 0,
-        "operating_life": 0.0,
-        "overstress": 0,
-        "package_id": 0,
-        "power_operating": 0.0,
-        "power_rated": 0.0,
-        "power_ratio": 0.0,
-        "reason": "",
-        "resistance": 0.0,
-        "specification_id": 0,
-        "technology_id": 0,
-        "temperature_active": 35.0,
-        "temperature_case": 0.0,
-        "temperature_dormant": 25.0,
-        "temperature_hot_spot": 0.0,
-        "temperature_junction": 0.0,
-        "temperature_knee": 25.0,
-        "temperature_rated_max": 0.0,
-        "temperature_rated_min": 0.0,
-        "temperature_rise": 0.0,
-        "theta_jc": 0.0,
-        "type_id": 0,
-        "voltage_ac_operating": 0.0,
-        "voltage_dc_operating": 0.0,
-        "voltage_esd": 0.0,
-        "voltage_rated": 0.0,
-        "voltage_ratio": 0.0,
-        "weight": 0.0,
-        "years_in_production": 1,
-    }
-
-
 @pytest.fixture(scope="class")
 def test_tablemodel(test_program_dao):
     """Get a data manager instance for each test class."""
